agents_os_v3/requirements.txt
# AOS v3 API + tooling (Team 61 GATE_0). Install: pip install -r agents_os_v3/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
apscheduler>=3.10.0
psycopg2-binary>=2.9.9
PyYAML>=6.0.1
python-ulid>=2.2.0
jsonschema>=4.19.0
pytest>=7.4.0
eval_type_backport>=0.2.0
httpx>=0.27.0
